✨ adds delete recipient feature
